Change "Pipeline" folder's name to "rossml".
