+++ rossml/Metrics.csv
@@ -1,0 +1,6 @@
+,Metric
+MAE,0.005
+MSE,0.0
+R2,0.983
+R2_adj,0.982
+explained variance,0.983
