+++ rossml/img/qq_plot_cxx.png
