+++ rossml/img/qq_plot_cyx.png
