+++ rossml/img/qq_plot_cyy.png
