+++ rossml/img/qq_plot_kxx.png
