+++ rossml/img/qq_plot_kxy.png
