+++ rossml/img/qq_plot_kyx.png
