+++ rossml/img/qq_plot_kyy.png
