+++ rossml/img/qq_plot_speeds.png
