+++ rossml/img/standardized_error_cxx_plot.png
