+++ rossml/img/standardized_error_cxy_plot.png
