+++ rossml/img/standardized_error_cyx_plot.png
