+++ rossml/img/standardized_error_cyy_plot.png
