+++ rossml/img/standardized_error_kxx_plot.png
