+++ rossml/img/standardized_error_kxy_plot.png
