+++ rossml/img/standardized_error_kyx_plot.png
