+++ rossml/img/standardized_error_kyy_plot.png
